+++ oalg-abg/CHANGELOG.md
@@ -63,3 +63,10 @@ and this project adheres to the
 
 ### Adaption
 - adaption to oalg-base-3.0.0.0
+
+## [2.0.0.2] - 2025-10-13
+
+### Added
+- Functions for conversion between Z-Vectors and abelian elements: abges, abgevec, vecabge
+- Instance OrientdOpl for AbElement.
+- Functions abgSomeFree, abgxs